+++ test/helpers/key_value.js
@@ -6,7 +6,7 @@ import { Helper } from './_helper'
 import { isUndefined } from 'util';
 
 export class KeyValue extends Helper {
-  async getEntryValue(key) {
+  async getEntryValue (key) {
     try {
       const { data } = await this.sdk.horizon.keyValue.get(key)
       switch (data.type.value) {
@@ -17,7 +17,7 @@ export class KeyValue extends Helper {
         case xdr.KeyValueEntryType.string().value:
           return data.stringValue
         default:
-          throw new Error("Unexpected key value entry type: " + data.type.name)
+          throw new Error('Unexpected key value entry type: ' + data.type.name)
       }
     } catch (e) {
       if (e instanceof NotFoundError) {
@@ -28,7 +28,20 @@ export class KeyValue extends Helper {
     }
   }
 
-  putEntries(entries) {
+  putUint64Entries (entries) {
+    const operations = Object
+      .entries(entries)
+      .map(([key, value]) => {
+        let entryType = KEY_VALUE_ENTRY_TYPES.uint64
+
+        return { key, value, entryType }
+      })
+      .map(opts => ManageKeyValueBuilder.putKeyValue(opts))
+
+    return this.submit(operations)
+  }
+
+  putEntries (entries) {
     const operations = Object
       .entries(entries)
       .map(([key, value]) => {
@@ -53,7 +66,7 @@ export class KeyValue extends Helper {
     return this.submit(operations)
   }
 
-  delete(key) {
+  delete (key) {
     return this.submit(ManageKeyValueBuilder.deleteKeyValue({ key }))
   }
 }