+++ CHANGELOG.md
@@ -6,12 +6,17 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 Please check our [developers guide](https://gitlab.com/tokend/developers-guide)
 for further information about branching and tagging conventions.
+
 ## [Unreleased]
 ### Fixed
 - A bug with the omitted key chain seeds if the `seed` property present
 
 ### Changed
 - Now default signer can be provided on the wallet creation
+
+## [1.14.0-rc.3] - 2020-09-04
+## Added
+- Create, update, remove data reviewable requests
 
 ## [1.14.0-rc.2] - 2020-08-07
 ### Fixed
@@ -315,7 +320,8 @@ for further information about branching and tagging conventions.
 
 ## [1.3.0] - 2019-03-01
 
-[Unreleased]: https://github.com/tokend/new-js-sdk/compare/1.14.0-rc.2...HEAD
+[Unreleased]: https://github.com/tokend/new-js-sdk/compare/1.14.0-rc.3...HEAD
+[1.14.0-rc.3]: https://github.com/tokend/new-js-sdk/compare/1.14.0-rc.2...1.14.0-rc.3
 [1.14.0-rc.2]: https://github.com/tokend/new-js-sdk/compare/1.14.0-rc.1...1.14.0-rc.2
 [1.14.0-rc.1]: https://github.com/tokend/new-js-sdk/compare/1.14.0-rc.0...1.14.0-rc.1
 [1.14.0-rc.0]: https://github.com/tokend/new-js-sdk/compare/1.13.1...1.14.0-rc.0